+++ test/.eslintrc.cjs
@@ -6,5 +6,9 @@ module.exports = {
     node: true,
     mocha: true
   },
+  parserOptions: {
+    // this is required for dynamic import()
+    ecmaVersion: 2020
+  },
   ignorePatterns: ['node_modules/']
 };